Fixed compilation warnings found in some compilers. https://tracker.physiomeproject.org/show_bug.cgi?id=3163

+++ core/src/string_const.cpp
@@ -180,7 +180,7 @@ char *strdupS( const char *str )
 
 const string getDirectory( const string filename )
 {
-    unsigned int index;
+		size_t index;
     
     index = filename.rfind( NIX_PATH_SEP );
 #ifdef WIN32
